Fixed padding miscalculation in ColorTexture Fixed integer overflow error in FloatSimpleIO Fixed soft lock in designer grid renderer math Added support for font scaling in text boxes Added preliminary methods for new themed GUI hooks

+++ src/main/java/betterquesting/api2/utils/EntityPlayerPreview.java
@@ -27,7 +27,7 @@ public class EntityPlayerPreview extends EntityOtherPlayerMP
 	{
 		super(worldIn, gameProfileIn);
 		this.resource = new ResourceLocation(BetterQuesting.MODID, "textures/skin_cache/" + gameProfileIn.getName());
-        this.getDataManager().set(PLAYER_MODEL_FLAG, Byte.valueOf((byte)1));
+        this.getDataManager().set(PLAYER_MODEL_FLAG, (byte)1);
 	}
 	
 	@Override
